make graph_lower a submodule
